requirements.txt
discord.py[voice]
python-dotenv
openai
requests
aiohttp
